# crates/birdnet-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BirdNet-Behavior HTTP API",
    "version": "0.16.0",
    "summary": "JSON API for a BirdNet-Behavior acoustic monitoring station: open read-only endpoints, plus a small token-authenticated write surface.",
...
    "license": {
      "name": "CC BY-NC-SA 4.0",
      "url": "https://github.com/tomtom215/BirdNet-Behavior/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "/api/v2",
      "description": "Versioned API root, relative to the station's base URL (default http://localhost:8502)."
    }
  ],
  "security": [],
...
  "paths": {
...
    "/detections": {
      "get": {
        "tags": [
          "detections"
        ],
        "summary": "List detections",
        "description": "Paginated detection log, newest first. When `limit`/`offset` are supplied the response also carries `total_count`.",
        "operationId": "listDetections",
        "parameters": [
          {
            "$ref": "#/components/parameters/Date"
          },
          {
            "$ref": "#/components/parameters/Species"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of detections.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid `date` (expected YYYY-MM-DD).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/species/detail": {
      "get": {
        "tags": [
          "species"
        ],
        "summary": "Per-species detail",
        "description": "Summary statistics plus hourly activity for one species.",
        "operationId": "speciesDetail",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Common or scientific name of the species.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Species detail.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "species": {
                      "$ref": "#/components/schemas/SpeciesSummary"
                    },
                    "hourly_activity": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/HourlyCount"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such species.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/species/image/{scientific_name}/file": {
      "get": {
        "tags": [
          "species"
        ],
        "summary": "Species image bytes",
        "description": "The cached image itself. Checks a configured custom image directory first, then the Wikipedia cache.",
        "operationId": "speciesImageFile",
        "parameters": [
          {
            "$ref": "#/components/parameters/ScientificName"
          }
        ],
        "responses": {
          "200": {
            "description": "Image bytes.",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "No cached image available."
          }
        }
      }
    },
...
    "/recordings/{filename}": {
      "get": {
        "tags": [
          "recordings"
        ],
        "summary": "Stream a recording",
        "description": "Streams a single audio file, honouring HTTP `Range` requests (so an `<audio>` element can seek). The filename is the basename only; path separators and `..` are rejected.",
        "operationId": "serveRecording",
        "parameters": [
          {
            "$ref": "#/components/parameters/Filename"
          }
        ],
        "responses": {
          "200": {
            "description": "The full audio file.",
            "content": {
              "audio/wav": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "audio/mpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "audio/flac": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "206": {
            "description": "A requested byte range (Range request)."
          },
          "400": {
            "description": "Invalid filename."
          },
          "404": {
            "description": "Recording not found."
          },
          "416": {
            "description": "Requested range not satisfiable."
          }
        }
      }
    },
    "/spectrogram/{filename}": {
      "get": {
        "tags": [
          "recordings"
        ],
        "summary": "Spectrogram PNG for a recording",
        "description": "Renders (and caches) a mel spectrogram PNG of a recording. Optional label parameters overlay the species, confidence and time.",
        "operationId": "serveSpectrogram",
        "parameters": [
          {
            "$ref": "#/components/parameters/Filename"
          },
          {
            "name": "species",
            "in": "query",
            "required": false,
            "description": "Overlay species label.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "confidence",
            "in": "query",
            "required": false,
            "description": "Overlay confidence (0-100).",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": false,
            "description": "Overlay time label.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Spectrogram image.",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filename."
          },
          "404": {
            "description": "Recording not found."
          },
          "422": {
            "description": "Could not render a spectrogram for this recording."
          },
          "503": {
            "description": "Renderer temporarily unavailable (too many concurrent renders)."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The station's `BNB_API_TOKEN`, resolved from its configuration file and then its environment. Required by every write endpoint, and by the settings read — a station's configuration is not public even with its credentials taken out. The read-only detection, species and analytics endpoints accept no credential and ignore one if it is sent."
      }
    },
    "parameters": {
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of results.",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 0,
          "default": 20
        }
      },
      "Offset": {
        "name": "offset",
        "in": "query",
        "required": false,
        "description": "Number of results to skip (pagination).",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 0,
          "default": 0
        }
      },
      "Date": {
        "name": "date",
        "in": "query",
        "required": false,
        "description": "Calendar day filter (YYYY-MM-DD).",
        "schema": {
          "type": "string",
          "format": "date"
        }
      },
      "Species": {
        "name": "species",
        "in": "query",
        "required": false,
        "description": "Filter to one species (common or scientific name).",
        "schema": {
          "type": "string"
        }
      },
...
      "ScientificName": {
        "name": "scientific_name",
        "in": "path",
        "required": true,
        "description": "Scientific name of the species.",
        "schema": {
          "type": "string"
        }
      }
    },
...
    "schemas": {
...
      "DetectionList": {
        "type": "object",
        "properties": {
          "detections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetectionRow"
            }
          },
          "count": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "total_count": {
            "type": "integer",
            "description": "Total matching rows (present on paginated responses)."
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        },
        "required": [
          "detections"
        ]
      },
...
      "SpeciesSummary": {
        "type": "object",
        "properties": {
          "com_name": {
            "type": "string"
          },
          "sci_name": {
            "type": "string"
          },
          "count": {
            "type": "integer"
          },
          "avg_confidence": {
            "type": "number",
            "format": "double"
          },
          "first_seen": {
            "type": "string",
            "format": "date"
          },
          "last_seen": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "HourlyCount": {
        "type": "object",
        "properties": {
          "hour": {
            "type": "string",
            "examples": [
              "06"
            ]
          },
          "count": {
            "type": "integer"
          }
        },
        "required": [
          "hour",
          "count"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "examples": [
              "not found"
            ]
          },
          "path": {
            "type": "string",
            "examples": [
              "/api/v2/nope"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  }
}